Few validations for network.
Added few checks and validations for network customization. Hyper-V's
external switch is best suited for vagrant. Though Internal switch will
also be as good as external switch, but it requires a bit manualy
configuration from the user.
Hemce customization will automatically create only an external switch is
not exist, the rest it will check if the switch is created and let the
user create any other swith other than external.